Changed set duty to range from -400 to 400.

+++ Drivers/t200/t200wrapper.h
@@ -4,6 +4,6 @@
 // TIM2 | Channel 1,2,3,4: Output Compare No Output | Prescaler: 15 if the chip clock is 16mhz
 // TIM3 | Channel 1,2: Output Compare No Output | Prescaler: 15 if the chip clock is 16mhz
 void Thruster_Init();
-void t200_setspeed(TIM_TypeDef *TIMx, uint8_t channel, uint32_t speed);
+void t200_setspeed(TIM_TypeDef *TIMx, uint8_t channel, int16_t speed);
 
 #endif /* INC_T200WRAPPER_H_ */
